requirements.txt
torch>=2.1.0
transformers>=4.44.0
peft>=0.12.0
trl>=0.9.0
bitsandbytes>=0.43.0
datasets>=2.20.0
accelerate>=0.33.0
wandb>=0.17.0
scikit-learn>=1.5.0
sentencepiece>=0.2.0
protobuf>=5.27.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
